--- CRUD_Alumnos/crud_alumnos/src/main/java/com/mycompany/crud_alumnos/Alumno.java
@@ -1,30 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package com.mycompany.crud_alumnos;
-
-/**
- *
- * @author gabo
- */
-public class Alumno {
-    String carnet;
-    String nombreCompleto;
-    
-    public String getCarnet() {
-        return carnet;
-    }
-    
-    public void setCarnet(String carnet) {
-        this.carnet = carnet;
-    }
-    
-    public String getNombreCompleto() {
-        return nombreCompleto;
-    }
-    
-    public void setNombreCompleto(String nombreCompleto) {
-        this.nombreCompleto = nombreCompleto;
-    }
-}
